+++ TokanPages.ClientApp/src/Api/Paths/index.ts
