+++ TokanPages.ClientApp/src/Shared/Components/ArticleCard/articleCard.tsx
@@ -5,7 +5,7 @@ import { ApplicationState } from "../../../Store/Configuration";
 import { ArticleSelectionAction } from "../../../Store/Actions";
 import { GetShortText } from "../../../Shared/Services/Utilities";
 import { MapLanguageId } from "../../../Shared/Services/languageService";
-import { ARTICLE_PATH, GET_ARTICLE_MAIN_IMAGE_URL } from "../../../Api/Request";
+import { ARTICLE_PATH, GET_ARTICLE_MAIN_IMAGE_URL } from "../../../Api";
 import { ArticleCardView } from "./View/articleCardView";
 
 interface ArticleCardProps {
