+++ TokanPages.ClientApp/src/Store/Actions/Articles/articleInfo.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../Configuration";
 import { ArticleItemBase } from "../../../Shared/Components/RenderContent/Models";
-import { DispatchExecuteAction, ExecuteRequest, GET_ARTICLE_INFO } from "../../../Api/Request";
+import { DispatchExecuteAction, ExecuteRequest, GET_ARTICLE_INFO } from "../../../Api";
 
 export const REQUEST = "REQUEST_ARTICLE_INFO";
 export const RECEIVE = "RECEIVE_ARTICLE_INFO";
