+++ TokanPages.ClientApp/src/Store/Actions/Content/contentPageData.ts
@@ -1,7 +1,7 @@
 import { ApplicationAction } from "../../Configuration";
 import { RequestPageDataResultDto } from "../../../Api/Models";
 import { GetVerifiedComponents } from "../../../Shared/Services/Utilities";
-import { DispatchExecuteAction, ExecuteRequest, REQUEST_PAGE_DATA } from "../../../Api/Request";
+import { DispatchExecuteAction, ExecuteRequest, REQUEST_PAGE_DATA } from "../../../Api";
 
 export const CLEAR = "CLEAR_PAGE_DATA";
 export const REQUEST = "REQUEST_PAGE_DATA";
