+++ TokanPages.ClientApp/src/Store/Actions/Newsletters/newsletterUpdate.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../Configuration";
 import { UpdateNewsletterDto } from "../../../Api/Models";
-import { DispatchExecuteAction, ExecuteRequest, UPDATE_NEWSLETTER } from "../../../Api/Request";
+import { DispatchExecuteAction, ExecuteRequest, UPDATE_NEWSLETTER } from "../../../Api";
 
 export const UPDATE = "UPDATE_SUBSCRIBER";
 export const RESPONSE = "UPDATE_SUBSCRIBER_RESPONSE";
