+++ TokanPages.ClientApp/src/Store/Actions/Users/Crud/userNoteRead.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../../Configuration";
 import { UserNoteDto, UserNoteResultDto } from "../../../../Api/Models";
-import { DispatchExecuteAction, ExecuteRequest, GET_USER_NOTE } from "../../../../Api/Request";
+import { DispatchExecuteAction, ExecuteRequest, GET_USER_NOTE } from "../../../../Api";
 
 export const RECEIVE = "GET_USER_NOTE_RECEIVE";
 export const RESPONSE = "GET_USER_NOTE_RESPONSE";
