+++ TokanPages.ClientApp/src/Store/Actions/Users/Crud/userNotesRead.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../../Configuration";
 import { UserNotesDto, UserNotesResultDto } from "../../../../Api/Models";
-import { DispatchExecuteAction, ExecuteRequest, GET_USER_NOTES } from "../../../../Api/Request";
+import { DispatchExecuteAction, ExecuteRequest, GET_USER_NOTES } from "../../../../Api";
 
 export const RECEIVE = "GET_USER_NOTES_RECEIVE";
 export const RESPONSE = "GET_USER_NOTES_RESPONSE";
