+++ TokanPages.ClientApp/src/Store/Actions/Users/userEmailVerification.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../Configuration";
 import { VerifyUserEmailDto } from "../../../Api/Models";
-import { VERIFY_USER_EMAIL, DispatchExecuteAction, ExecuteRequest } from "../../../Api/Request";
+import { VERIFY_USER_EMAIL, DispatchExecuteAction, ExecuteRequest } from "../../../Api";
 
 export const VERIFY = "VERIFY_USER_EMAIL";
 export const CLEAR = "VERIFY_USER_EMAIL_CLEAR";
