+++ TokanPages.ClientApp/src/Store/Actions/Users/userNotification.ts
@@ -1,6 +1,6 @@
 import { ApplicationAction } from "../../Configuration";
 import { NotificationData, NotificationRequest } from "../../../Api/Models";
-import { NOTIFY_WEB_URL, DispatchExecuteAction, ExecuteRequest } from "../../../Api/Request";
+import { NOTIFY_WEB_URL, DispatchExecuteAction, ExecuteRequest } from "../../../Api";
 
 export const CLEAR = "NOTIFICATION_CLEAR";
 export const NOTIFY = "NOTIFICATION_NOTIFY";
